# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Fast minimalist URL shortener",
        "title": "Scissors",
        "contact": {
            "name": "Fuad Olatunji",
            "url": "https://fuadolatunji.me",
            "email": "fuadolatunji@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/analytics/{id}": {
            "get": {
                "description": "Fetch analytics of a shortened URL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Short URL analytics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The URL ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.JSONResponse-model_URL"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.JSONResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.JSONResponse-any"
                        }
                    }
                }
            }
        },
        "/api/health": {
            "get": {
                "description": "Healty status checker for load balancers and monitoring systems.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Healty status checker",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.JSONResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.JSONResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.JSONResponse-any"
                        }
                    }
                }
            }
        },
        "/api/shorten": {
            "post": {
                "description": "Shorten a URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Shorten a URL",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.SBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.JSONResponse-controller_SRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.JSONResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.JSONResponse-any"
                        }
                    }
                }
            }
        },
        "/api/unpublish/{id}": {
            "delete": {
                "description": "Unpublish/Delete a shortened URL.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Unpublish/Delete a shortened URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The URL ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.JSONResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.JSONResponse-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.JSONResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.JSONResponse-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.JSONResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "controller.JSONResponse-controller_SRes": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/controller.SRes"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "controller.JSONResponse-model_URL": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.URL"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "controller.SBody": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "controller.SRes": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "url": {
                    "description": "shortened URL",
                    "type": "string"
                }
            }
        },
        "model.URL": {
            "type": "object",
            "properties": {
                "clicks": {
                    "type": "integer"
                },
                "country_metrics": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}